Add files via upload
This  project focuses on analyzing demographic data from the CIA World Factbook using SQL in Jupyter Notebook. The database includes information such as population size, growth rate, and land/water area of each country on Earth. The database can be downloaded from the project's website, and we'll use SQL queries and Python code to clean, explore, and analyze the data. Throughout the project, we'll answer questions about the data, such as which countries have the highest/lowest population growth rates or which countries have the highest population densities. By the end of the project, we'll have gained  a deeper understanding of global demographic trends.